+++ include/core/dcp/model/pdu/DcpPdu.hpp
@@ -136,7 +136,7 @@ public:
 
     void setPduSize(size_t pduSize) {
         this->stream_size = pduSize + PDU_LENGTH_INDICATOR_SIZE;
-        *((uint32_t*) this->stream) = pduSize;
+        *((uint32_t*) this->stream) = static_cast<uint32_t>(pduSize);
     }
 
 protected:
